Moved generateURL token to lucia package
