+++ packages/lucia/package.json
@@ -14,6 +14,7 @@
     "@ei/drizzle": "*",
     "@lucia-auth/adapter-postgresql": "^2.0.1",
     "@lucia-auth/oauth": "^3.3.0",
+    "drizzle-orm": "^0.28.6",
     "ip": "^1.1.8",
     "lucia": "^2.7.1",
     "nanoid": "^5.0.1"
