+++ packages/lucia/tsconfig.json
@@ -4,7 +4,6 @@
   "exclude": ["node_modules"],
   "compilerOptions": {
     "module": "CommonJS",
-    "isolatedModules": true,
-    "baseUrl": "."
+    "isolatedModules": true
   }
 }
